init to fload

+++ test_01.rb
@@ -1,4 +1,4 @@
-g = ARGV[0].to_i
+g = ARGV[0].to_f
 r = ARGV[1].to_i
 f = Math.sqrt(2*g*r).round(2)
